treewide: Fix invalid bit shift operations.
UB Sanitizer reports:
  tests/test-hash.c:59:40:
  runtime error: shift exponent 64 is too large for 64-bit type
  'long unsigned int'
      0 0x44c3c9 in get_range128 tests/test-hash.c:59
      1 0x44cb2e in check_hash_bytes128 tests/test-hash.c:178
      2 0x44d14d in test_hash_main tests/test-hash.c:282
      [...]
  ofproto/ofproto-dpif-xlate.c:5607:45:
  runtime error: left shift of 65535 by 16 places cannot be represented
  in type 'int'
      0 0x53fe9f in xlate_sample_action ofproto/ofproto-dpif-xlate.c:5607
      1 0x54d625 in do_xlate_actions ofproto/ofproto-dpif-xlate.c:7160
      2 0x553b76 in xlate_actions ofproto/ofproto-dpif-xlate.c:7806
      3 0x4fcb49 in upcall_xlate ofproto/ofproto-dpif-upcall.c:1237
      4 0x4fe02f in process_upcall ofproto/ofproto-dpif-upcall.c:1456
      5 0x4fda99 in upcall_cb ofproto/ofproto-dpif-upcall.c:1358
      [...]
  tests/test-util.c:89:23:
  runtime error: left shift of 1 by 31 places cannot be represented in
  type 'int'
      0 0x476415 in test_ctz tests/test-util.c:89
      [...]
  lib/dpif-netlink.c:396:33:
  runtime error: left shift of 1 by 31 places cannot be represented in
  type 'int'
      0 0x571b9f in dpif_netlink_open lib/dpif-netlink.c:396

Acked-by: Aaron Conole <aconole@redhat.com>
Acked-by: Paolo Valerio <pvalerio@redhat.com>

+++ ofproto/ofproto-dpif-xlate.c
@@ -5653,7 +5653,8 @@ xlate_sample_action(struct xlate_ctx *ctx,
 
     /* Scale the probability from 16-bit to 32-bit while representing
      * the same percentage. */
-    uint32_t probability = (os->probability << 16) | os->probability;
+    uint32_t probability =
+        ((uint32_t) os->probability << 16) | os->probability;
 
     /* If ofp_port in flow sample action is equel to ofp_port,
      * this sample action is a input port action. */
